Moved item_list and item_detail to just item.py
